refacotring

+++ ReadWriteMemory/Models/MemoryAddress.cs
@@ -9,8 +9,8 @@
 /// <code>MemoryAddress memoryAddress = new(0x1234567, 0x42, 0x420, 0x69)</code>
 /// <code>MemoryAddress memoryAddress = new(0x1234567)</code>
 /// </example>
-/// <para>See <seealso cref="MemoryAddress(long, int[])"/></para> 
-/// See <seealso cref="MemoryAddress(long, string, int[])"/>
+/// <para>See <seealso cref="MemoryAddress(nuint, int[])"/></para> 
+/// See <seealso cref="MemoryAddress(nuint, string, int[])"/>
 /// </summary>
 public sealed record MemoryAddress
 {
@@ -25,7 +25,7 @@ public sealed record MemoryAddress
     /// <param name="address"></param>
     /// <param name="moduleName"></param>
     /// <param name="offsets"></param>
-    public MemoryAddress(long address, string moduleName = "", params int[]? offsets)
+    public MemoryAddress(nuint address, string moduleName = "", params int[]? offsets)
     {
         Address = address;
         ModuleName = moduleName.ToLower();
@@ -43,14 +43,14 @@ public sealed record MemoryAddress
     /// </summary>
     /// <param name="address"></param>
     /// <param name="offsets"></param>
-    public MemoryAddress(long address, params int[] offsets)
+    public MemoryAddress(nuint address, params int[] offsets)
     {
         Address = address;
         ModuleName = string.Empty;
         Offsets = offsets;
     }
 
-    internal long Address { get; }
+    internal nuint Address { get; }
 
     internal string ModuleName { get; }
